// frontend/src/api/model/githubCom4H1RZooraInternalDomainPracticeRoomView.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainClass } from './githubCom4H1RZooraInternalDomainClass';
import type { GithubCom4H1RZooraInternalDomainClassSession } from './githubCom4H1RZooraInternalDomainClassSession';
import type { GithubCom4H1RZooraInternalDomainPracticeRoomStats } from './githubCom4H1RZooraInternalDomainPracticeRoomStats';
import type { GithubCom4H1RZooraInternalDomainPracticeSubmission } from './githubCom4H1RZooraInternalDomainPracticeSubmission';
import type { GithubCom4H1RZooraInternalDomainUser } from './githubCom4H1RZooraInternalDomainUser';

export interface GithubCom4H1RZooraInternalDomainPracticeRoomView {
  attachments?: string[];
  can_grade?: boolean;
  can_submit?: boolean;
  class?: GithubCom4H1RZooraInternalDomainClass;
  class_id?: string;
  class_session?: GithubCom4H1RZooraInternalDomainClassSession;
  class_session_id?: string;
  content?: string;
  created_at?: string;
  end_time?: string;
  id?: string;
  max_score?: number;
  mine?: boolean;
  my_submission?: GithubCom4H1RZooraInternalDomainPracticeSubmission;
  organization_id?: string;
  start_time?: string;
  /** present only when CanGrade */
  stats?: GithubCom4H1RZooraInternalDomainPracticeRoomStats;
  /** derived student status */
  status?: string;
  title?: string;
  updated_at?: string;
  user?: GithubCom4H1RZooraInternalDomainUser;
  user_id?: string;
}